image viewer script 작성 및 코드 개선
image viewer 와 table view script 분리
