+++ Task1/Assets/Scripts/DicomImageViewer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6fdfd21c57a2bc408f6f0ba18055314
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
