+++ Task1/Assets/Scripts/DicomStudyRow.cs
@@ -1,9 +1,6 @@
 using UnityEngine;
 using UnityEngine.UI;
-using System.Collections.Generic;
-using Newtonsoft.Json.Linq;
 using Mars.db;
-using Unity.VisualScripting;
 namespace StudyRow
 {
     public class DicomStudyRow : MonoBehaviour
