+++ Task1/Assets/Scripts/DicomTableview.cs
@@ -8,11 +8,9 @@ using UnityEngine.EventSystems;
 using StudyRow;
 using Newtonsoft.Json;
 using System.IO;
-using Mars;
-using System.Linq;
-using UnityEditor;
+using SeriesImageViewer;
 
-public class DicomImageViewer : MonoBehaviour
+public class DicomTableView : MonoBehaviour
 {
     [SerializeField] private ScrollRect studyScrollview;
     [SerializeField] private ScrollRect seriesScrollview;
@@ -29,8 +27,7 @@ public class DicomImageViewer : MonoBehaviour
     [SerializeField] private InputField searchInputField;
     [SerializeField] private Text searchText;
 
-    [SerializeField] private RawImage volumeImage;
-    [SerializeField] private Slider volumeSlider;
+    private DicomImageViewer imageViewer;
 
     private const string dicomURL = "http://10.10.20.173:5080/v2/Dicom/";
     private const string dicomVolumeURL = "http://10.10.20.173:5080/dicom/";
@@ -41,13 +38,10 @@ public class DicomImageViewer : MonoBehaviour
     void OnClickStudyRow()
     {
         GameObject studyRowContent = EventSystem.current.currentSelectedGameObject.transform.parent.gameObject;
-
-        // 클릭된 버튼의 부모 object (study data row) 를 key, 해당 row 의 studyId 를 value
         string studyId = dicomStudyRowContentsId[studyRowContent];
         RemoveSeriesObject();
         StartCoroutine(GetSeriesData(studyId));
         SetStudyInVisible(studyId);
-
     }
 
     void OnClickReset()
@@ -72,57 +66,13 @@ public class DicomImageViewer : MonoBehaviour
 
     void OnClickSeriesData()
     {
-        volumeImage.texture = null;
-        volumeImage.gameObject.SetActive(true);
-        volumeSlider.gameObject.SetActive(true);
+        imageViewer.ResetImageAndSlider();
         Button seriesDataButton = EventSystem.current.currentSelectedGameObject.GetComponent<Button>();
         string seriesId = dicomSeriesButtonsData[seriesDataButton][0];
         string volumePath = dicomSeriesButtonsData[seriesDataButton][1];
         string volumeFile = Application.persistentDataPath + "/" + seriesId + ".nrrd";
-        if (!File.Exists(volumeFile)) StartCoroutine(GetVolumeData(seriesId, volumeFile, volumePath));
-        else ShowVolumeImage(volumeFile);
-    }
-
-    async void ShowVolumeImage(string volumeFile)
-    {
-        NrrdRaw nrrdData = await NrrdRaw.LoadAsync(volumeFile);
-        NrrdHeader nrrdHeader = nrrdData.Header;
-        int width = nrrdHeader.sizes[0];
-        int height = nrrdHeader.sizes[1];
-        int slices = nrrdHeader.sizes[2];
-
-        short[] imageArray = nrrdData.Int16Data.ToArray();
-        short minValue = imageArray.Min();
-        short maxValue = imageArray.Max();
-        float[] normalizedArray = imageArray.Select(value => (value - minValue) / (float)(maxValue - minValue)).ToArray();
-
-        Dictionary<int, Color[]> slicedArrayDict = new();
-        for (int i = 0; i < slices; i++)
-        {
-            float[] slicedImageArray = normalizedArray.Skip(width * height * i).Take(width * height).ToArray();
-            Color[] slicedColors = slicedImageArray.Select(x => new Color(x, x, x)).ToArray();
-            slicedArrayDict[i] = slicedColors;
-        }
-
-        // slider 의 시작 지점이 항상 index 0 이 되도록 설정
-        volumeSlider.value = 0;
-        volumeSlider.maxValue = slices - 1; // slice 의 index 가 0부터 시작했기 때문에 -1
-
-        Texture2D volumeTexture = new(width, height);
-        volumeTexture.SetPixels(slicedArrayDict[0]);
-        volumeTexture.Apply();
-        volumeImage.texture = volumeTexture;
-
-        // slider 에 새로운 데이터를 사용하기위해 초기화
-        volumeSlider.onValueChanged.RemoveAllListeners();
-        volumeSlider.onValueChanged.AddListener((value) => OnSliderValueChanged(value, volumeTexture, slicedArrayDict));
-    }
-
-    void OnSliderValueChanged(float sliderValue, Texture2D volumeTexture, Dictionary<int, Color[]> slicedColorDict)
-    {
-        volumeTexture.SetPixels(slicedColorDict[(int)sliderValue]);
-        volumeTexture.Apply();
-        volumeImage.texture = volumeTexture;
+        if (!File.Exists(volumeFile)) StartCoroutine(imageViewer.GetVolumeData(seriesId, volumeFile, volumePath, dicomVolumeURL));
+        else imageViewer.ShowVolumeImage(volumeFile);
     }
 
     // 데이터의 크기만큼 화면이 출력되도록
@@ -143,11 +93,10 @@ public class DicomImageViewer : MonoBehaviour
 
     void SetStudyVisible()
     {
-        // study data가 활성화되면 series 데이터와 Volume 이미지는 비활성화되어 화면이 겹치는 것 방지
+        // study data가 활성화되면 series 데이터와 volume image 는 비활성화되어 화면이 겹치는 것 방지
         seriesScrollview.gameObject.SetActive(false);
         searchContent.SetActive(true);
-        volumeImage.gameObject.SetActive(false);
-        volumeSlider.gameObject.SetActive(false);
+        imageViewer.SetVolumeVisible(false);
         resetButton.gameObject.SetActive(false);
 
         foreach (string rowStudyId in dicomStudyIdRowContents.Keys)
@@ -203,6 +152,8 @@ public class DicomImageViewer : MonoBehaviour
     void Start()
     {
         StartCoroutine(GetStudyData());
+        resetButton.onClick.AddListener(OnClickReset);
+        imageViewer = FindObjectOfType<DicomImageViewer>();
     }
 
     IEnumerator GetStudyData()
@@ -223,7 +174,6 @@ public class DicomImageViewer : MonoBehaviour
                 AddDicomStudyRow(studyData);
             }
             searchButton.onClick.AddListener(() => OnClickSearch(dicomStudyList));
-            resetButton.onClick.AddListener(OnClickReset);
             ResetScrollview();
         }
     }
@@ -246,25 +196,6 @@ public class DicomImageViewer : MonoBehaviour
                 AddDicomSeriesRow(seriesData);
             }
             seriesButton.gameObject.SetActive(false);
-        }
-    }
-
-    IEnumerator GetVolumeData(string seriesId, string volumeFile, string volumePath)
-    {
-        string volumeURLPath = dicomVolumeURL + volumePath;
-
-        UnityWebRequest reqVolume = UnityWebRequest.Get(volumeURLPath);
-        yield return reqVolume.SendWebRequest();
-
-        if (reqVolume.result != UnityWebRequest.Result.Success)
-        {
-            Debug.Log($"Dicom Series ID {seriesId} 의 Volume 파일은 존재하지 않습니다.");
-            volumeSlider.maxValue = 0;
-        }
-        else
-        {
-            File.WriteAllBytes(volumeFile, reqVolume.downloadHandler.data);
-            ShowVolumeImage(volumeFile);
         }
     }
 }
