+++ Task1/Assets/Scripts/References.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 371cdb61da8b875498f8158a2b3ec197
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
